fix: directly interpolate gen.py path into buildPhase
On aarch64-linux, there is an updateAutotoolsGnuConfigScriptsPhase which
also uses a variable called $script. As a result, Python tries to
execute something like /build/config.guess and dies instead of executing
the gen.py script as intended. Correct this issue by not storing the
script path in a variable at all.